tests/wpt/tests/html/browsers/history/the-history-interface/pushstate-replacestate-empty-string/replacestate.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>history.replaceState() with an empty string URL</title>
<link rel="help" href="https://github.com/whatwg/html/issues/9343">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {
  location.hash = "test";

  history.replaceState(null, null, "");
  assert_equals(location.hash, "#test");
});
</script>
